requirements.txt
Django==4.1.5
openpyxl==3.0.10
djangorestframework==3.14.0
tqdm==4.64.1
pandas==1.5.3
mysqlclient==2.1.1
flatten-json==0.1.13
django-mysql==4.8.0
pytest-django==4.5.2
